add cast
git-svn-id: https://svn.r-project.org/R/trunk@60794 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/modules/X11/devX11.c
@@ -2672,7 +2672,7 @@ static Rboolean X11_Locator(double *x, double *y, pDevDesc dd)
 static int translate_key(KeySym keysym)
 {
     if ((keysym >= XK_F1) && (keysym <= XK_F12))
-        return knF1 + keysym - XK_F1;
+        return knF1 + (int)keysym - XK_F1;
     else
     {
         switch (keysym)
